Ajuste de metodo toBigDecimal para a classe Utils
