+++ grails-app/services/com/mini/asaas/payment/PaymentAdapter.groovy
@@ -1,6 +1,7 @@
 package com.mini.asaas.payment
 
 import com.mini.asaas.utils.FormatUtils
+import com.mini.asaas.utils.Utils
 import com.mini.asaas.utils.enums.PaymentStatus
 import com.mini.asaas.utils.enums.PaymentType
 
@@ -16,7 +17,7 @@ class PaymentAdapter {
 
     public PaymentAdapter(Map params) {
         this.paymentType = params.paymentType
-        this.value = FormatUtils.toBigDecimal(params.value)
+        this.value = Utils.toBigDecimal(params.value)
         this.status = PaymentStatus.AWAITING_PAYMENT
         this.dueDate = FormatUtils.toDate(params.dueDate, "yyyy-MM-dd")
     }
